% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dum9.R
\name{dum9}
\alias{dum9}
\title{dum9}
\usage{
dum9(x)
}
\arguments{
\item{x}{a list or data.frame}
}
\value{
returns the list/df but with return
}
\description{
Takes output from histXCII and tells you the return
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 lapply(dum9,LGNN)
 }
}
}
