% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetChain_R.R
\name{GetChain_R}
\alias{GetChain_R}
\title{GetChain_R}
\usage{
GetChain_R(symbol, option_type)
}
\arguments{
\item{symbol}{The stock assets market symbol, ex: "AAPL"}

\item{option_type}{The option type, ("call" or "put")}
}
\value{
Options Chain for given input
}
\description{
sources GetDasChain.py to globalenv, then runs the given stock symbol and option type
}
\details{
DETAILS
}
\examples{
GetChain_R("AAPL","call")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[reticulate]{source_python}}
}
