% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histXCI.R
\name{histXCI}
\alias{histXCI}
\title{FUNCTION_TITLE}
\usage{
histXCI(file)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_sub}},\code{\link[stringr]{str_split}},\code{\link[stringr]{str_remove}},\code{\link[stringr]{str_locate}},\code{\link[stringr]{str_replace}}
}
