% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LGNN_Agg}
\alias{LGNN_Agg}
\title{LargestNet_Noon Database}
\format{An object of class \code{data.frame} with 222 rows and 14 columns.}
\usage{
LGNN_Agg
}
\description{
Largest NET gainers/losers taken at noon. Consists of every existing date of data thus far.
}
\keyword{datasets}
